Re-organised folders and files structure, and appropriately renamed some files.
