+++ database/models/analytics_model.py
