+++ database/models/server_model.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 from beanie import Document, Link
 from pydantic import BaseModel, Field
 
-from models.user_model import User
+from database.models.user_model import User
 
 
 class Channel(BaseModel):
