+++ utils/stats_utils.py
@@ -2,12 +2,12 @@ from datetime import datetime
 
 from beanie.odm.operators.update.array import AddToSet, Pull
 from beanie.odm.operators.update.general import Set
-from bot_globals import calculate_scores, client, logger
-from models.server_model import Rankings, Server, UserRank
-from models.user_model import History, Submissions, User
 
-from utils.leaderboards import get_score
-from utils.questions import get_problems_solved_and_rank
+from bot_globals import calculate_scores, client, logger
+from database.models.server_model import Rankings, Server, UserRank
+from database.models.user_model import History, Submissions, User
+from utils.leaderboards_utils import get_score
+from utils.questions_utils import get_problems_solved_and_rank
 
 
 async def update_rankings(server: Server, now: datetime, timeframe: str) -> None:
