+++ utils/views_utils.py
@@ -5,7 +5,7 @@ import discord
 from embeds.channels_embeds import channel_remove_embed, channel_set_embed
 from embeds.general_embeds import help_embed, not_creator_embed
 from embeds.misc_embeds import error_embed
-from utils.channels import get_options, save_channel_options
+from utils.channels_utils import get_options, save_channel_options
 
 
 class Pagination(discord.ui.View):
